# OpenAPI.yaml
openapi: 3.0.0
info:
  title: Eurostat Data API
  description: API for searching and retrieving data from Eurostat
  version: 1.0.0

servers:
  - url: https://statchat.onrender.com

paths:
  /privacy:
    get:
      operationId: getPrivacyPolicy
      summary: Privacy Policy
      description: Returns the privacy policy page.
      responses:
        200:
          description: Privacy policy page
          content:
            text/html:
              schema:
                type: string

  /search_for_tables:
    post:
      operationId: searchForTables
      summary: Search Eurostat Tables
      description: Searches for tables in Eurostat that match the user's question.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                user_question:
                  type: string
                  description: The question the user is asking.
              required:
                - user_question
      responses:
        200:
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResults'

  /get_table_variables:
    post:
      operationId: getTableVariables
      summary: Get Table Variables
      description: Gets the variables for a table in Eurostat.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                table_code:
                  type: string
                  description: Code of the Eurostat table.
              required:
                - table_code
      responses:
        200:
          description: Variables for the table
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableVariables'

  /get_table_data:
    post:
      operationId: getData
      summary: Get Data for Table
      description: >
        Retrieves table data based on query parameters as JSON. The query keys 
        correspond to variable IDs from `get_table_variables`. The response is 
        a CSV string with column headers as variables.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                table_code:
                  type: string
                  description: Code of the Eurostat table.
                query:
                  type: object
                  description: >
                    JSON object containing parameters for the query. 
                    The keys are variable IDs/codes corresponding to those provided 
                    by the `get_table_variables` endpoint. For non specified variables, 
                    all values are returned.
                  additionalProperties:
                    type: array
                    items:
                      type: string
              required:
                - table_code
      responses:
        200:
          description: CSV data for the specified table
          content:
            text/csv:
              schema:
                type: string
                description: CSV formatted string. The first row contains column headers (variables).

components:
  schemas:
    SearchResults:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/SearchResult'

    SearchResult:
      type: object
      properties:
        code:
          type: string
        title:
          type: string
        date_start:
          type: integer
        date_end:
          type: integer

    TableVariables:
      type: object
      description: Variables available for a specific table.
      properties:
        variable_name:
          type: object
          description: The structure of each variable.
          additionalProperties: 
            type: object
            properties:
              text:
                type: string
              values:
                type: object
                additionalProperties:
                  type: string

    TableData:
      type: string
      description: >
        A string representing CSV formatted data. The first row of the CSV 
        contains the column headers, which correspond to the variables from the query.